requirements.txt

pandas>=2.2.2
numpy>=1.26.4
scikit-learn>=1.5.1
xgboost>=2.0.3
SQLAlchemy>=2.0.31
PyMySQL>=1.1.1
python-dotenv>=1.0.1
joblib>=1.4.2
